requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==35.0.0
execnet==1.9.0
h11==0.12.0
idna==3.3
iniconfig==1.1.1
numpy==1.21.3
outcome==1.1.0
packaging==21.0
Pillow==8.4.0
pluggy==1.0.0
py==1.10.0
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==3.0.3
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==2.4.0
requests==2.26.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
webdriver-manager==3.5.1
wsproto==1.0.0
